requirements.txt
asgiref==3.5.0
boto3==1.21.22
botocore==1.24.22
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-ninja==0.17.0
django-storages==1.12.3
django-taggit==2.1.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
jmespath==1.0.0
mysqlclient==2.1.0
Pillow==9.0.1
pydantic==1.9.0
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
uvicorn==0.17.6
wincertstore==0.2